Abstract the base functionality of StringToItemParser
